tái cấu trúc #2
