+++ events/client/error.js
@@ -1,5 +1,4 @@
 const { Events } = require("discord.js");
-const config = require("../config");
 
 module.exports = {
 	name: Events.Error,
