+++ events/client/interactionCreate.js
@@ -1,5 +1,5 @@
 const { Events, CommandInteraction, PermissionsBitField } = require("discord.js");
-const config = require("./../config");
+const config = require("../../config");
 const { useCooldowns, useCommands, useFunctions } = require("@zibot/zihooks");
 
 const Cooldowns = useCooldowns();
