+++ events/player/playerSkip.js
@@ -1,0 +1,17 @@
+const { EmbedBuilder } = require("discord.js");
+const { GuildQueueEvent } = require("discord-player");
+
+module.exports = {
+	name: GuildQueueEvent.playerSkip,
+	type: "Player",
+	execute: async (queue, track) => {
+		const embed = new EmbedBuilder()
+			.setDescription(`Đã Skip: [${track?.title}](${track?.url}) \`[${track?.duration}]\``)
+			.setThumbnail(track?.thumbnail)
+			.setColor("Random");
+		const replied = await queue.metadata?.channel?.send({ embeds: [embed], fetchReply: true }).catch((e) => {});
+		setTimeout(function () {
+			replied.delete().catch((e) => console.log(e));
+		}, 5000);
+	},
+};
